# Swagger.yaml
openapi: 3.1.0
info:
  title: Auto Escola 3ESA
  description: API REST para gerenciamento de instruções da Auto Escola 3ESA
  contact:
    name: Auto Escola 3ESA
    email: autoescola3esa@email.com.br
  license:
    name: Apache 2.0
    url: http://springdoc.org
servers:
  - url: http://localhost:8085
    description: Generated server url
paths:
  /instrutores:
    get:
      tags:
        - instrutor-controller
      operationId: listarInstrutores
      parameters:
        - name: page
          in: query
          description: Zero-based page index (0..N)
          required: false
          schema:
            type: integer
            default: 0
            minimum: 0
        - name: size
          in: query
          description: The size of the page to be returned
          required: false
          schema:
            type: integer
            default: 10
            minimum: 1
        - name: sort
          in: query
          description: "Sorting criteria in the format: property,(asc|desc). Default sort
            order is ascending. Multiple sort criteria are supported."
          required: false
          schema:
            type: array
            default:
              - nome,ASC
            items:
              type: string
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/PageDadosListagemInstrutor"
      security:
        - bearer-key: []
    put:
      tags:
        - instrutor-controller
      operationId: atualizarInstrutor
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DadosAtualizacaoInstrutor"
        required: true
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/DadosDetalhamentoInstrutor"
      security:
        - bearer-key: []
    post:
      tags:
        - instrutor-controller
      operationId: cadastrarInstrutor
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DadosCadastroInstrutor"
        required: true
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/DadosDetalhamentoInstrutor"
      security:
        - bearer-key: []
  /alunos:
    get:
      tags:
        - aluno-controller
      operationId: listarAlunos
      parameters:
        - name: paginacao
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/Pageable"
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/PageDadosListagemAluno"
    put:
      tags:
        - aluno-controller
      operationId: atualizarAluno
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DadosAtualizacaoAluno"
        required: true
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/DadosDetalhamentoAluno"
    post:
      tags:
        - aluno-controller
      operationId: cadastrarAluno
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DadosCadastroAluno"
        required: true
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/DadosDetalhamentoAluno"
  /login:
    post:
      tags:
        - login-controller
      operationId: efetuarLogin
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DadosLogin"
        required: true
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/DadosTokenJWT"
  /instrucoes:
    post:
      tags:
        - instrucao-controller
      operationId: agendarInstrucao
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DadosAgendamentoInstrucao"
        required: true
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/DadosDetalhamentoInstrucao"
  /instrucoes/{id}/cancelamento:
    patch:
      tags:
        - instrucao-controller
      operationId: cancelarInstrucao
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DadosCancelamentoInstrucao"
        required: true
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/DadosDetalhamentoInstrucao"
  /instrutores/{id}:
    get:
      tags:
        - instrutor-controller
      operationId: detalharInstrutor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/DadosDetalhamentoInstrutor"
      security:
        - bearer-key: []
    delete:
      tags:
        - instrutor-controller
      operationId: excluirInstrutor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
      security:
        - bearer-key: []
  /health_check:
    get:
      tags:
        - health-check-controller
      operationId: healthCheck
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                type: string
  /alunos/{id}:
    get:
      tags:
        - aluno-controller
      operationId: detalharAlunos
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/DadosDetalhamentoAluno"
    delete:
      tags:
        - aluno-controller
      operationId: excluirAluno
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
components:
  schemas:
    DadosAtualizacaoInstrutor:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nome:
          type: string
        email:
          type: string
        telefone:
          type: string
        especialidade:
          type: string
          enum:
            - MOTOS
            - CARROS
            - VANS
            - CAMINHOES
        endereco:
          $ref: "#/components/schemas/DadosEndereco"
      required:
        - id
    DadosEndereco:
      type: object
      properties:
        logradouro:
          type: string
          minLength: 1
        numero:
          type: string
        complemento:
          type: string
        bairro:
          type: string
          minLength: 1
        cidade:
          type: string
          minLength: 1
        uf:
          type: string
          minLength: 1
          pattern: "[A-Z]{2}"
        cep:
          type: string
          minLength: 1
          pattern: \d{8}
      required:
        - bairro
        - cep
        - cidade
        - logradouro
        - uf
    DadosDetalhamentoInstrutor:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nome:
          type: string
        email:
          type: string
        telefone:
          type: string
        cnh:
          type: string
        especialidade:
          type: string
          enum:
            - MOTOS
            - CARROS
            - VANS
            - CAMINHOES
        endereco:
          $ref: "#/components/schemas/DadosEndereco"
        ativo:
          type: boolean
    DadosAtualizacaoAluno:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nome:
          type: string
        email:
          type: string
        telefone:
          type: string
        endereco:
          $ref: "#/components/schemas/DadosEndereco"
      required:
        - id
    DadosDetalhamentoAluno:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nome:
          type: string
        email:
          type: string
        telefone:
          type: string
        cpf:
          type: string
        endereco:
          $ref: "#/components/schemas/DadosEndereco"
        ativo:
          type: boolean
    DadosLogin:
      type: object
      properties:
        login:
          type: string
          minLength: 1
        senha:
          type: string
          minLength: 1
      required:
        - login
        - senha
    DadosTokenJWT:
      type: object
      properties:
        tokenJWT:
          type: string
    DadosCadastroInstrutor:
      type: object
      properties:
        nome:
          type: string
          minLength: 1
        email:
          type: string
          format: email
          minLength: 1
        telefone:
          type: string
          minLength: 1
        cnh:
          type: string
          minLength: 1
          pattern: "[0-9]{9,11}"
        especialidade:
          type: string
          enum:
            - MOTOS
            - CARROS
            - VANS
            - CAMINHOES
        endereco:
          $ref: "#/components/schemas/DadosEndereco"
      required:
        - cnh
        - email
        - especialidade
        - nome
        - telefone
    DadosAgendamentoInstrucao:
      type: object
      properties:
        idAluno:
          type: integer
          format: int64
        idInstrutor:
          type: integer
          format: int64
        especialidade:
          type: string
          enum:
            - MOTOS
            - CARROS
            - VANS
            - CAMINHOES
        dataHora:
          type: string
          format: date-time
      required:
        - dataHora
        - especialidade
        - idAluno
    DadosDetalhamentoInstrucao:
      type: object
      properties:
        id:
          type: integer
          format: int64
        idAluno:
          type: integer
          format: int64
        nomeAluno:
          type: string
        idInstrutor:
          type: integer
          format: int64
        nomeInstrutor:
          type: string
        especialidade:
          type: string
        dataHora:
          type: string
          format: date-time
        status:
          type: string
        motivoCancelamento:
          type: string
        dataCancelamento:
          type: string
          format: date-time
    DadosCadastroAluno:
      type: object
      properties:
        nome:
          type: string
          minLength: 1
        email:
          type: string
          format: email
          minLength: 1
        telefone:
          type: string
          minLength: 1
        cpf:
          type: string
          minLength: 1
          pattern: "[0-9]{11}"
        endereco:
          $ref: "#/components/schemas/DadosEndereco"
      required:
        - cpf
        - email
        - nome
        - telefone
    DadosCancelamentoInstrucao:
      type: object
      properties:
        motivoCancelamento:
          type: string
          enum:
            - ALUNO_DESISIU
            - INSTRUTOR_CANCELOU
            - OUTROS
      required:
        - motivoCancelamento
    DadosListagemInstrutor:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nome:
          type: string
        email:
          type: string
        especialidade:
          type: string
          enum:
            - MOTOS
            - CARROS
            - VANS
            - CAMINHOES
    PageDadosListagemInstrutor:
      type: object
      properties:
        totalPages:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: "#/components/schemas/DadosListagemInstrutor"
        number:
          type: integer
          format: int32
        sort:
          $ref: "#/components/schemas/SortObject"
        pageable:
          $ref: "#/components/schemas/PageableObject"
        numberOfElements:
          type: integer
          format: int32
        first:
          type: boolean
        last:
          type: boolean
        empty:
          type: boolean
    PageableObject:
      type: object
      properties:
        offset:
          type: integer
          format: int64
        sort:
          $ref: "#/components/schemas/SortObject"
        paged:
          type: boolean
        pageSize:
          type: integer
          format: int32
        pageNumber:
          type: integer
          format: int32
        unpaged:
          type: boolean
    SortObject:
      type: object
      properties:
        empty:
          type: boolean
        sorted:
          type: boolean
        unsorted:
          type: boolean
    Pageable:
      type: object
      properties:
        page:
          type: integer
          format: int32
          minimum: 0
        size:
          type: integer
          format: int32
          minimum: 1
        sort:
          type: array
          items:
            type: string
    DadosListagemAluno:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nome:
          type: string
        email:
          type: string
    PageDadosListagemAluno:
      type: object
      properties:
        totalPages:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: "#/components/schemas/DadosListagemAluno"
        number:
          type: integer
          format: int32
        sort:
          $ref: "#/components/schemas/SortObject"
        pageable:
          $ref: "#/components/schemas/PageableObject"
        numberOfElements:
          type: integer
          format: int32
        first:
          type: boolean
        last:
          type: boolean
        empty:
          type: boolean
  securitySchemes:
    bearer-key:
      type: http
      scheme: bearer
      bearerFormat: JWT

  